Modernization
Controls options updated to become Modifiers. Listens for Alt, Ctrl, Shift, or Super interchangeably. Needed to do this as modern macs do not have a right Ctrl.

Restructured main.go into a cmd folder. Recuddled the code.

Removed rand.Seed(...) as that is automatic in Go 1.20.

Updated flags using library.

Removed whatever the fps timer was attempting to do.

Resaves the config if a value was updated. Specifically, if a trigger was loaded with the value "controls" the config will be updated and resaved with the values "modifiers"