Resizable!
LimitedWindow's interface is now a struct: GameWindow. It handles the responsibilities of the canvas and the window such as responding to a resize or getting input. Maps a fixed size canvas to whatever the window size is. Allows for useless sizes. That's on you.

New DrawMouse function on the GameWindow that was super useful in mapping input during resizing.