Configurable Window Size
Specify the window size in the config. THIS ALSO DECIDES THE STARTING SIZE OF THE BUFFER. For the record, this is where all the hard work of placing the UI elements relatively based on the width and height of the canvas made making the window resizable CRIMINALLY easy.

Config validation. If something is screwy, revert to defaults and save. You broke it, I'll fix it. Also, leaning into the validator. Custom checks and corrections removed in favor of defaulting.

Simplified resizing by including it in the Draw func removing unnecessary GameLoop call.